requirements/base.txt
django==2.2.3
djangorestframework==3.10.1
markdown==3.1.1
django-filter==2.2.0
logzero==1.5.0
